aoc/src/test/java/com/yaodanzhang/aoc/year2022/Day10Test.java
package com.yaodanzhang.aoc.year2022;

import static com.yaodanzhang.aoc.input.Inputs.inputOf;
import static org.assertj.core.api.Assertions.assertThat;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

class Day10Test {

  private Day10 puzzle;

  @BeforeEach
  void setUp() {
    puzzle = new Day10(inputOf(this.getClass()));
  }

  @Test
  void should_pass_part_1() {
    assertThat(puzzle.part1Result()).isEqualTo(13140);
  }

  @Test
  void should_pass_part_2() {
    assertThat(puzzle.part2Result()).isEqualTo(0);
  }

  @Test
  void should_print_out_result() {
    Day10 realPuzzle = new Day10(inputOf(Day10.class));
    System.out.println(realPuzzle.part1Result());
    System.out.println(realPuzzle.part2Result());
  }
}